+++ myqueue/test/workflow_test.py
@@ -11,12 +11,31 @@ from myqueue.test.hello import workflow as hello
 
 def test_flow1(mq):
     script = Path(__file__).with_name('flow1.py')
+
     mq(f'workflow {script}')
     mq.wait()
     assert mq.states() == 'ddddddd'
-    mq(f'workflow {script}')
+
+mq(f'workflow {script}')
     mq.wait()
     assert mq.states() == 'dddddddd'
+
+    mq('rm -sd')
+    mq.wait()
+    assert mq.states() == 'dddddddd'
+
+    mq('rm -sd --force')
+    mq.wait()
+    assert mq.states() == ''
+
+    mq(f'workflow {script}')
+    mq.wait()
+    assert mq.states() == ''
+
+    Path('builtins.print.result').unlink()
+    mq(f'workflow {script}')
+    mq.wait()
+    assert mq.states() == 'd'
 
 
 def test_direct_cached_flow1(tmp_path, capsys):